Move test utils into utils dir
