+++ __test__/pages/Home.test.tsx
@@ -1,5 +1,5 @@
 import React from "react"
-import { render, screen } from "../../__test__/test-utils"
+import { render, screen } from "../../src/utils/testUtils"
 
 import Home from "../../pages/index"
