+++ __test__/pages/test-pickup-times/index.test.tsx
@@ -1,5 +1,5 @@
 import React from "react"
-import { screen, render } from "../../../__test__/test-utils"
+import { screen, render } from "../../../src/utils/testUtils"
 
 import TestPickupTimes from "../../../pages/test-pickup-times/index"
