+++ src/utils/testUtils.tsx
@@ -1,6 +1,6 @@
 import React, { type ReactElement } from "react"
 import { render, type RenderOptions } from "@testing-library/react"
-import { FeedbackProvider } from "../src/context/FeedbackContext"
+import { FeedbackProvider } from "../context/FeedbackContext"
 
 const AllTheProviders = ({ children }: { children: React.ReactNode }) => {
   return <FeedbackProvider value={null}>{children}</FeedbackProvider>
